Renamed to normalized_worm.py for consistency
Now all module names are in lowercase
