+++ movement_validation/user_config_example.txt
@@ -8,7 +8,6 @@
   
 """
 
-#EXAMPLE_DATA_PATH = '/Users/jameshokanson/Google Drive/OpenWorm/OpenWorm Public/movement_validation/example_data'
 EXAMPLE_DATA_PATH = r'C:\Users\RNEL\Google Drive\OpenWorm\OpenWorm Public\movement_validation\example_data'
